+++ src/ImboClient/ImboClient.php
@@ -350,6 +350,36 @@ class ImboClient extends GuzzleClient {
     }
 
     /**
+     * Generate a short URL
+     *
+     * @param Http\ImageUrl $imageUrl An instance of an imageUrl
+     * @return
+     * @throws
+     */
+    public function generateShortUrl(Http\ImageUrl $imageUrl) {
+        $transformations = $imageUrl->getTransformations();
+
+        if ($transformations) {
+            $transformations = '?t[]=' . implode('&t[]=', $transformations);
+        } else {
+            $transformations = null;
+        }
+
+        $params = array(
+            'publicKey' => $this->getConfig('publicKey'),
+            'imageIdentifier' => $imageUrl->getImageIdentifier(),
+            'extension' => $imageUrl->getExtension(),
+            'query' => $transformations,
+        );
+
+        return $this->getCommand('GenerateShortUrl', array(
+            'publicKey' => $this->getConfig('publicKey'),
+            'imageIdentifier' => $imageUrl->getImageIdentifier(),
+            'params' => json_encode($params),
+        ))->execute();
+    }
+
+    /**
      * Get all server URL's
      *
      * @return string[]
@@ -432,28 +462,6 @@ class ImboClient extends GuzzleClient {
         );
 
         return Http\MetadataUrl::factory($url, $this->getConfig('privateKey'));
-    }
-
-    /**
-     * Get the short URL of an image (with optional transformations)
-     *
-     * @param Http\ImageUrl $imageUrl An instance of an imageUrl
-     * @param boolean $asString Set to true to return the URL as a string
-     * @return GuzzleUrl|string
-     * @throws InvalidArgumentException
-     */
-    public function getShortUrl(Http\ImageUrl $imageUrl, $asString = false) {
-        try {
-            $shortUrl = (string) $this->head((string) $imageUrl)->send()->getHeader('x-imbo-shorturl');
-
-            if (!$asString) {
-                $shortUrl = GuzzleUrl::factory($shortUrl);
-            }
-        } catch (GuzzleException $e) {
-            throw new InvalidArgumentException('Could not fetch image properties for image: ' . $imageUrl);
-        }
-
-        return $shortUrl;
     }
 
     /**